2、数据集成/数据标准 优化完成

+++ src/main/java/com/bigdata/backstage/modules/homepage/vo/DataOverviewVo.java
@@ -28,5 +28,5 @@ public class DataOverviewVo {
     /**
      * 数据总记录数
      */
-    private Integer rowCount;
+    private Long rowCount;
 }
